refactor(domain): organize src into directory-per-aggregate
Group multi-file domain aggregates into their own directories (team,
artifact, weapon, character, profile) while keeping single-file shared
primitives (iso-timestamp, uuid, problem-detail) at the src root. This
mirrors the one-concern-per-directory layout used in apps/api/src and
improves discoverability as aggregates grew validation and
representation files.

The representations/ layer stays as a separate top-level directory
rather than co-locating with aggregates, keeping the serialization
concern grouped. The index.ts barrel re-exports are unchanged, so
apps/api and apps/web need no import changes.

Closes #589
