+++ packages/domain/src/artifact/artifact-plan-validation.test.ts
