+++ packages/domain/src/artifact/artifact-plan-validation.ts
