+++ packages/domain/src/character/collection-character.test.ts
@@ -3,13 +3,13 @@
 
 import { describe, expect, it } from 'vitest';
 
+import type { ISOTimestamp } from '../iso-timestamp.js';
 import {
   assertCollectionCharacter,
   isValidConstellationLevel,
   MAX_CONSTELLATION_LEVEL,
   MIN_CONSTELLATION_LEVEL,
 } from './collection-character.js';
-import type { ISOTimestamp } from './iso-timestamp.js';
 
 const VALID_TIMESTAMP = '2024-01-15T12:00:00Z' as ISOTimestamp;
