+++ packages/domain/src/index.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: MIT
 
 export { issue, isValid, prefixPaths, type ValidationIssue } from '@genshin/validation';
-export { type ArtifactPlan } from './artifact-plan.js';
-export { validateArtifactPlan } from './artifact-plan-validation.js';
-export type { AuthIdentity } from './auth-identity.js';
+export { type ArtifactPlan } from './artifact/artifact-plan.js';
+export { validateArtifactPlan } from './artifact/artifact-plan-validation.js';
+export type { AuthIdentity } from './profile/auth-identity.js';
 export {
   assertCollectionCharacter,
   isValidConstellationLevel,
@@ -12,7 +12,7 @@ export {
   MIN_CONSTELLATION_LEVEL,
   type CharacterId,
   type CollectionCharacter,
-} from './collection-character.js';
+} from './character/collection-character.js';
 export {
   assertCollectionTeam,
   createEmptyTeam,
@@ -27,8 +27,8 @@ export {
   type CollectionTeam,
   type CollectionTeamMembers,
   type TeamSlot,
-} from './collection-team.js';
-export { type CollectionTeamMember } from './collection-team-member.js';
+} from './team/collection-team.js';
+export { type CollectionTeamMember } from './team/collection-team-member.js';
 export {
   assertCollectionWeapon,
   isValidRefinementLevel,
@@ -36,7 +36,7 @@ export {
   MIN_REFINEMENT_LEVEL,
   type CollectionWeapon,
   type CollectionWeaponId,
-} from './collection-weapon.js';
+} from './weapon/collection-weapon.js';
 export { isISOTimestamp, nowTimestamp, type ISOTimestamp } from './iso-timestamp.js';
 export type { ProblemDetail } from './problem-detail.js';
 export {
@@ -65,6 +65,6 @@ export {
   type ProfileResponse,
 } from './representations/json/profile.js';
 
-export { validateTeam, validateTeams, type TeamValidationContext } from './team-validation.js';
-export { assertUserProfile, type ProfileUpdate, type UserProfile } from './user-profile.js';
+export { validateTeam, validateTeams, type TeamValidationContext } from './team/team-validation.js';
+export { assertUserProfile, type ProfileUpdate, type UserProfile } from './profile/user-profile.js';
 export type { UUID } from './uuid.js';
