+++ packages/domain/src/representations/collection-json/teams.test.ts
@@ -4,8 +4,12 @@
 import { describe, expect, it } from 'vitest';
 
 import { deserialiseTeam, serialiseTeam } from './teams.js';
-import type { CollectionTeam, CollectionTeamMembers, TeamSlot } from '../../collection-team.js';
 import type { ISOTimestamp } from '../../iso-timestamp.js';
+import type {
+  CollectionTeam,
+  CollectionTeamMembers,
+  TeamSlot,
+} from '../../team/collection-team.js';
 import type { UUID } from '../../uuid.js';
 
 const BASE_URL = 'http://localhost:8080';
