+++ packages/domain/src/representations/collection-json/teams.ts
@@ -21,10 +21,10 @@ import {
   type Template,
 } from '@genshin/collection-json';
 
-import type { ArtifactPlan } from '../../artifact-plan.js';
-import type { CollectionTeamMember } from '../../collection-team-member.js';
-import type { CollectionTeam, CollectionTeamMembers } from '../../collection-team.js';
-import { assertCollectionTeam, MAX_TEAM_MEMBERS } from '../../collection-team.js';
+import type { ArtifactPlan } from '../../artifact/artifact-plan.js';
+import type { CollectionTeamMember } from '../../team/collection-team-member.js';
+import type { CollectionTeam, CollectionTeamMembers } from '../../team/collection-team.js';
+import { assertCollectionTeam, MAX_TEAM_MEMBERS } from '../../team/collection-team.js';
 
 const TEAM_TEMPLATE: Template = {
   data: [
