+++ packages/domain/src/representations/collection-json/weapons.test.ts
@@ -4,9 +4,9 @@
 import { describe, expect, it } from 'vitest';
 
 import { deserialiseWeapon, serialiseWeapon } from './weapons.js';
-import type { CollectionWeapon } from '../../collection-weapon.js';
 import type { ISOTimestamp } from '../../iso-timestamp.js';
 import type { UUID } from '../../uuid.js';
+import type { CollectionWeapon } from '../../weapon/collection-weapon.js';
 
 const BASE_URL = 'http://localhost:8080';
 const VALID_TIMESTAMP = '2024-01-15T12:00:00Z' as ISOTimestamp;
