+++ packages/domain/src/team/team-validation.test.ts
@@ -3,11 +3,11 @@
 
 import { describe, expect, it } from 'vitest';
 
-import type { ArtifactPlan } from './artifact-plan.js';
+import type { ArtifactPlan } from '../artifact/artifact-plan.js';
+import type { UUID } from '../uuid.js';
 import type { CollectionTeamMembers, TeamSlot } from './collection-team.js';
 import type { TeamValidationContext } from './team-validation.js';
 import { validateTeam, validateTeams } from './team-validation.js';
-import type { UUID } from './uuid.js';
 
 const EMPTY_MEMBERS: CollectionTeamMembers = [null, null, null, null];
