+++ packages/domain/src/weapon/collection-weapon.ts
@@ -4,9 +4,9 @@
 import type { Weapon } from '@genshin/game-data';
 import { getWeaponById } from '@genshin/game-data';
 
-import type { ISOTimestamp } from './iso-timestamp.js';
-import { isISOTimestamp } from './iso-timestamp.js';
-import type { UUID } from './uuid.js';
+import type { ISOTimestamp } from '../iso-timestamp.js';
+import { isISOTimestamp } from '../iso-timestamp.js';
+import type { UUID } from '../uuid.js';
 
 export const MIN_REFINEMENT_LEVEL = 1;
 export const MAX_REFINEMENT_LEVEL = 5;
